+++ tencentcloud/service_tencentcloud_clb.go
@@ -3,10 +3,13 @@ package tencentcloud
 import (
 	"context"
 	"fmt"
-	"log"
-
+	"github.com/hashicorp/terraform/helper/resource"
+	"github.com/hashicorp/terraform/helper/schema"
 	clb "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/clb/v20180317"
 	"github.com/terraform-providers/terraform-provider-tencentcloud/tencentcloud/connectivity"
+	"log"
+	"strings"
+	"time"
 )
 
 type ClbService struct {
@@ -89,6 +92,7 @@ func (me *ClbService) DescribeLoadBalancerByFilter(ctx context.Context, params m
 }
 
 func (me *ClbService) DeleteLoadBalancerById(ctx context.Context, clbId string) error {
+
 	logId := GetLogId(ctx)
 	request := clb.NewDeleteLoadBalancerRequest()
 	request.LoadBalancerIds = []*string{&clbId}
@@ -100,7 +104,263 @@ func (me *ClbService) DeleteLoadBalancerById(ctx context.Context, clbId string) 
 	}
 	log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n",
 		logId, request.GetAction(), request.ToJsonString(), response.ToJsonString())
+	request_id := *response.Response.RequestId
+	retryErr := retrySet(request_id, me.client.UseClbClient())
+	if retryErr != nil {
+		return retryErr
+	}
 	return nil
+}
+
+func (me *ClbService) DescribeListenerById(ctx context.Context, id string) (clbListener *clb.Listener, errRet error) {
+	logId := GetLogId(ctx)
+	request := clb.NewDescribeListenersRequest()
+	items := strings.Split(id, "#")
+	if len(items) != 2 {
+		errRet = fmt.Errorf("id of resource.tencentcloud_clb_listener is wrong")
+		return
+	}
+
+	listenerId := items[0]
+	clbId := items[1]
+	request.ListenerIds = []*string{&listenerId}
+	request.LoadBalancerId = stringToPointer(clbId)
+
+	response, err := me.client.UseClbClient().DescribeListeners(request)
+	if err != nil {
+		log.Printf("[CRITAL]%s api[%s] fail, request body [%s], reason[%s]\n",
+			logId, request.GetAction(), request.ToJsonString(), err.Error())
+		errRet = err
+		return
+	}
+	log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n",
+		logId, request.GetAction(), request.ToJsonString(), response.ToJsonString())
+
+	if len(response.Response.Listeners) < 1 {
+		errRet = fmt.Errorf("Listener id is not found")
+		return
+	}
+	clbListener = response.Response.Listeners[0]
+	return
+}
+
+func (me *ClbService) DescribeListenersByFilter(ctx context.Context, params map[string]interface{}) (listeners []*clb.Listener, errRet error) {
+	logId := GetLogId(ctx)
+	request := clb.NewDescribeListenersRequest()
+	clbId := ""
+	for k, v := range params {
+		if k == "listener_id" {
+			items := strings.Split(v.(string), "#")
+			if len(items) != 2 {
+				errRet = fmt.Errorf("id of resource.tencentcloud_clb_listener is wrong")
+				return
+			}
+
+			listenerId := items[0]
+			clbId = items[1]
+			request.ListenerIds = []*string{stringToPointer(listenerId)}
+			request.LoadBalancerId = stringToPointer(clbId)
+		}
+		if k == "clb_id" {
+			if clbId == "" {
+				clbId = v.(string)
+				request.LoadBalancerId = stringToPointer(clbId)
+			}
+		}
+		if k == "protocol" {
+			request.Protocol = stringToPointer(v.(string))
+		}
+		if k == "port" {
+			port := int64(v.(int))
+			request.Port = &port
+		}
+
+	}
+
+	listeners = make([]*clb.Listener, 0)
+
+	response, err := me.client.UseClbClient().DescribeListeners(request)
+	if err != nil {
+		log.Printf("[CRITAL]%s api[%s] fail, request body [%s], reason[%s]\n",
+			logId, request.GetAction(), request.ToJsonString(), err.Error())
+		errRet = err
+		return
+	}
+	log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n",
+		logId, request.GetAction(), request.ToJsonString(), response.ToJsonString())
+
+	listeners = append(listeners, response.Response.Listeners...)
+
+	return
+}
+
+func (me *ClbService) DeleteListenerById(ctx context.Context, clbId string, listenerId string) error {
+	logId := GetLogId(ctx)
+	request := clb.NewDeleteListenerRequest()
+	request.ListenerId = stringToPointer(listenerId)
+	request.LoadBalancerId = stringToPointer(clbId)
+	response, err := me.client.UseClbClient().DeleteListener(request)
+	if err != nil {
+		log.Printf("[CRITAL]%s api[%s] fail, request body [%s], reason[%s]\n",
+			logId, request.GetAction(), request.ToJsonString(), err.Error())
+		return err
+	}
+	log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n",
+		logId, request.GetAction(), request.ToJsonString(), response.ToJsonString())
+	request_id := *response.Response.RequestId
+	retryErr := retrySet(request_id, me.client.UseClbClient())
+	if retryErr != nil {
+		return retryErr
+	}
+	return nil
+}
+
+func checkHealthCheckPara(ctx context.Context, d *schema.ResourceData, protocol string) (healthSetFlag bool, healthCheckPara *clb.HealthCheck, errRet error) {
+	var healthCheck clb.HealthCheck
+	healthSetFlag = false
+	healthCheckPara = &healthCheck
+	if v, ok := d.GetOk("health_check_switch"); ok {
+		healthSetFlag = true
+		vv := int64(v.(int))
+		healthCheck.HealthSwitch = &vv
+	}
+	if v, ok := d.GetOk("health_check_time_out"); ok {
+		healthSetFlag = true
+		vv := int64(v.(int))
+		healthCheck.TimeOut = &vv
+	}
+
+	if v, ok := d.GetOk("health_check_interval_time"); ok {
+		healthSetFlag = true
+		vv := int64(v.(int))
+		healthCheck.IntervalTime = &vv
+	}
+
+	if v, ok := d.GetOk("health_check_health_num"); ok {
+		healthSetFlag = true
+		vv := int64(v.(int))
+		healthCheck.HealthNum = &vv
+	}
+	if v, ok := d.GetOk("health_check_unhealth_num"); ok {
+		healthSetFlag = true
+		vv := int64(v.(int))
+		healthCheck.UnHealthNum = &vv
+	}
+
+	if v, ok := d.GetOk("health_check_http_code"); ok {
+		//仅适用于HTTP/HTTPS转发规则、TCP监听器的HTTP健康检查方式
+		if !(protocol == CLB_LISTENER_PROTOCOL_TCP) {
+			healthSetFlag = false
+			errRet = fmt.Errorf("health_check_http_code can only be set with protocol TCP %s", protocol)
+			return
+		} else {
+			healthSetFlag = true
+			vv := int64(v.(int))
+			healthCheck.HttpCode = &vv
+		}
+	}
+
+	if v, ok := d.GetOk("health_check_http_path"); ok {
+		if !(protocol == CLB_LISTENER_PROTOCOL_TCP) {
+			healthSetFlag = false
+			errRet = fmt.Errorf("health_check_http_path can only be set with protocol TCP")
+			return
+		} else {
+			healthSetFlag = true
+			healthCheck.HttpCheckPath = stringToPointer(v.(string))
+		}
+	}
+
+	if v, ok := d.GetOk("health_check_http_domain"); ok {
+		if !(protocol == CLB_LISTENER_PROTOCOL_TCP) {
+			healthSetFlag = false
+			errRet = fmt.Errorf("health_check_http_domain can only be set with protocol TCP")
+			return
+		} else {
+			healthSetFlag = true
+			healthCheck.HttpCheckDomain = stringToPointer(v.(string))
+		}
+	}
+
+	if v, ok := d.GetOk("health_check_http_method"); ok {
+		if !(protocol == CLB_LISTENER_PROTOCOL_TCP) {
+			healthSetFlag = false
+			errRet = fmt.Errorf("health_check_http_method can only be set with protocol TCP")
+			return
+		} else {
+			healthSetFlag = true
+			healthCheck.HttpCheckMethod = stringToPointer(v.(string))
+		}
+
+	}
+
+	if healthSetFlag == true {
+		if !(protocol == CLB_LISTENER_PROTOCOL_TCP || protocol == CLB_LISTENER_PROTOCOL_UDP) {
+			healthSetFlag = false
+			errRet = fmt.Errorf("health para can only be set with protocol TCP/UDP")
+			return
+		}
+		healthCheckPara = &healthCheck
+	}
+	return
+
+}
+
+func checkCertificateInputPara(ctx context.Context, d *schema.ResourceData) (certificateSetFlag bool, certPara *clb.CertificateInput, errRet error) {
+	certificateSetFlag = false
+	var certificateInput clb.CertificateInput
+	certificateSSLMode := ""
+	certificateId := ""
+	certificateCaId := ""
+
+	if v, ok := d.GetOk("certificate_ssl_mode"); ok {
+		certificateSetFlag = true
+		certificateSSLMode = v.(string)
+		certificateInput.SSLMode = stringToPointer(v.(string))
+	}
+
+	if v, ok := d.GetOk("certificate_id"); ok {
+		certificateSetFlag = true
+		certificateId = v.(string)
+		certificateInput.CertId = stringToPointer(v.(string))
+	}
+
+	if v, ok := d.GetOk("certificate_ca_id"); ok {
+		certificateSetFlag = true
+		certificateCaId = v.(string)
+		certificateInput.CertId = stringToPointer(v.(string))
+	}
+
+	if certificateSetFlag == true && certificateId == "" {
+		certificateSetFlag = false
+		errRet = fmt.Errorf("Certificate_key and certificate_name and certificate_content must be set when certificate_id is null")
+		return
+	}
+
+	if certificateSetFlag == true && certificateSSLMode == CERT_SSL_MODE_MUT && certificateCaId == "" {
+		certificateSetFlag = false
+		errRet = fmt.Errorf("Certificate_ca_key and certificate_ca_name and certificate_ca_content must be set when certificate_ca_id is null and ssl mode is 'MUTUAL' ")
+		return
+	}
+
+	certPara = &certificateInput
+
+	return
+}
+func retrySet(request_id string, meta *clb.Client) (err error) {
+	taskQueryRequest := clb.NewDescribeTaskStatusRequest()
+	taskQueryRequest.TaskId = &request_id
+	err = resource.Retry(10*time.Minute, func() *resource.RetryError {
+		taskResponse, e := meta.DescribeTaskStatus(taskQueryRequest)
+		if e != nil {
+			return resource.NonRetryableError(e)
+		}
+		if *taskResponse.Response.Status == int64(CLB_TASK_EXPANDING) {
+			return resource.RetryableError(fmt.Errorf("clb task status is %d, request_id is %s", *taskResponse.Response.Status, *taskResponse.Response.RequestId))
+		}
+		return nil
+	})
+	return
 }
 
 func flattenClbTagsMapping(tags []*clb.TagInfo) (mapping map[string]string) {